requirements.txt
absl-py==0.8.1
apache-beam==2.17.0
astunparse==1.6.3
attrs==19.3.0
avro-python3==1.9.2.1
awscli==1.16.303
backcall==0.1.0
bleach==3.1.5
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
crcmod==1.7
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
dill==0.3.0
docker==4.2.0
docopt==0.6.2
entrypoints==0.3
fastavro==0.21.24
fasteners==0.15
future==0.18.2
gast==0.3.3
google-api-core==1.17.0
google-api-python-client==1.8.4
google-apitools==0.5.28
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.3.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-cloud-storage==1.28.1
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.29.0
h5py==2.10.0
hdfs==2.5.8
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kfp==0.5.1
kfp-server-api==0.5.0
kubernetes==10.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
ml-metadata==0.21.2
mock==2.0.0
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==3.4.2
scikit-learn==0.21.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
strip-hints==0.1.9
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-data-validation==0.21.5
tensorflow-estimator==2.2.0
tensorflow-metadata==0.21.2
tensorflow-model-analysis==0.21.6
tensorflow-serving-api==2.1.0
tensorflow-transform==0.21.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tfx==0.21.4
tfx-bsl==0.21.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
